+++ docs/.vitepress/theme/components/FirstScene.vue
@@ -2,7 +2,7 @@
 import { ref, onMounted } from 'vue'
 import { BasicShadowMap, SRGBColorSpace, NoToneMapping } from 'three'
 
-import { TresCanvas } from '/@/tres'
+import { TresCanvas } from '@tresjs/core'
 import { OrbitControls } from '@tresjs/cientos'
 const LightRef = ref()