+++ src/core/nodeOps.ts
@@ -1,9 +1,9 @@
 import { RendererOptions } from 'vue'
-import { BufferAttribute, BufferGeometry, Material, Object3D } from 'three'
-import { useCamera, useLogger } from '/@/composables'
+import { BufferAttribute, BufferGeometry, Material } from 'three'
+import { useCamera, useLogger } from '../composables'
 import { isFunction } from '@alvarosabu/utils'
 import { catalogue } from './catalogue'
-import { TresInstance, TresObject } from '../types'
+import { EventHandlers, TresObject } from '../types'
 import { isHTMLTag, kebabToCamel } from '../utils'
 
 const onRE = /^on[^a-z]/
@@ -14,7 +14,7 @@ function noop(fn: string): any {
 }
 
 let fallback: TresObject | null = null
-
+let scene: Scene | null = null
 const OBJECT_3D_USER_DATA_KEYS = {
   GEOMETRY_VIA_PROP: 'tres__geometryViaProp',
   MATERIAL_VIA_PROP: 'tres__materialViaProp',
@@ -37,7 +37,7 @@ export const nodeOps: RendererOptions<TresObject, TresObject> = {
 
     if (tag === 'primitive') {
       if (props?.object === undefined) logError(`Tres primitives need a prop 'object'`)
-      const object = props.object as TresInstance
+      const object = props.object as TresObject
       name = object.type
       instance = Object.assign(object, { type: name, attach: props.attach, primitive: true })
     } else {
@@ -70,8 +70,8 @@ export const nodeOps: RendererOptions<TresObject, TresObject> = {
     const { GEOMETRY_VIA_PROP, MATERIAL_VIA_PROP } = OBJECT_3D_USER_DATA_KEYS
 
     if (instance.isObject3D) {
-      if (props?.material?.isMaterial) (instance as Object3D).userData[MATERIAL_VIA_PROP] = true
-      if (props?.geometry?.isBufferGeometry) (instance as Object3D).userData[GEOMETRY_VIA_PROP] = true
+      if (props?.material?.isMaterial) (instance as TresObject).userData[MATERIAL_VIA_PROP] = true
+      if (props?.geometry?.isBufferGeometry) (instance as TresObject).userData[GEOMETRY_VIA_PROP] = true
     }
 
     instance.events = {}
@@ -79,6 +79,7 @@ export const nodeOps: RendererOptions<TresObject, TresObject> = {
     return instance
   },
   insert(child, parent) {
+    if (parent && parent.isScene) scene = parent
     if (
       (child?.__vnode?.type === 'TresGroup' || child?.__vnode?.type === 'TresObject3D') &&
       parent === null &&
@@ -86,19 +87,19 @@ export const nodeOps: RendererOptions<TresObject, TresObject> = {
     ) {
       fallback = child
       return
+    } else if (child?.__vnode?.type.includes('Controls') && parent === null) {
+      fallback = scene
     }
 
     if (!parent) parent = fallback as TresObject
 
-    //vue core
-    /*  parent.insertBefore(child, anchor || null) */
     if (child?.isObject3D && parent?.isObject3D) {
       parent.add(child)
       child.dispatchEvent({ type: 'added' })
     } else if (child?.isFog) {
       parent.fog = child
     } else if (typeof child?.attach === 'string') {
-      child.__previousAttach = child[parent?.attach]
+      child.__previousAttach = child[parent?.attach as string]
       if (parent) {
         parent[child.attach] = child
       }
@@ -109,17 +110,17 @@ export const nodeOps: RendererOptions<TresObject, TresObject> = {
     // remove is only called on the node being removed and not on child nodes.
 
     if (node.isObject3D) {
-      const object3D = node as unknown as Object3D
+      const object3D = node as unknown as TresObject
 
-      const disposeMaterialsAndGeometries = (object3D: Object3D) => {
+      const disposeMaterialsAndGeometries = (object3D: TresObject) => {
         const { GEOMETRY_VIA_PROP, MATERIAL_VIA_PROP } = OBJECT_3D_USER_DATA_KEYS
 
-        if (!object3D.userData[MATERIAL_VIA_PROP]) (object3D as Object3D & { material: Material }).material?.dispose()
+        if (!object3D.userData[MATERIAL_VIA_PROP]) (object3D as TresObject & { material: Material }).material?.dispose()
         if (!object3D.userData[GEOMETRY_VIA_PROP])
-          (object3D as Object3D & { geometry: BufferGeometry }).geometry?.dispose()
+          (object3D as TresObject & { geometry: BufferGeometry }).geometry?.dispose()
       }
 
-      object3D.traverse(child => disposeMaterialsAndGeometries(child))
+      object3D.traverse((child: TresObject) => disposeMaterialsAndGeometries(child))
 
       disposeMaterialsAndGeometries(object3D)
     }
@@ -153,7 +154,8 @@ export const nodeOps: RendererOptions<TresObject, TresObject> = {
         if (!target?.set) root = chain.reduce((acc, key) => acc[kebabToCamel(key)], root)
       }
       if (isOn(key)) {
-        node.events[key] = nextValue
+        const eventHandlerKey: keyof EventHandlers = key as keyof EventHandlers // This is fine
+        node.events[eventHandlerKey] = nextValue
       }
       let value = nextValue
       if (value === '') value = true